LayoutTests/imported/w3c/web-platform-tests/event-timing/supported-types-consistent-with-self-expected.txt


FAIL supportedEntryTypes should be consistent with `'PerformanceEventTiming' in Self` assert_equals: expected true but got false
